requirements.txt
supabase
python-dotenv
pandas
requests
matplotlib
beautifulsoup4
re